+++ cashu-lib-common/src/test/java/xyz/tcheeric/cashu/common/VoucherPaymentRequestTest.java
@@ -3,6 +3,11 @@ package xyz.tcheeric.cashu.common;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import xyz.tcheeric.cashu.common.nut10.Nut10Option;
+import xyz.tcheeric.cashu.common.nut10.WellKnownSecret;
+import xyz.tcheeric.cashu.common.nut18.PaymentRequest;
+import xyz.tcheeric.cashu.common.nut18.VoucherPaymentRequest;
+import xyz.tcheeric.cashu.common.nut18.VoucherTransport;
 
 import java.util.List;