+++ cashu-lib-entities/src/main/java/xyz/tcheeric/cashu/entities/rest/nut04/PostMintQuoteBolt11Request.java
@@ -1,0 +1,11 @@
+package xyz.tcheeric.cashu.entities.rest.nut04;
+
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class PostMintQuoteBolt11Request extends PostMintQuoteRequest {
+
+    public PostMintQuoteBolt11Request(int amount) {
+        super(amount, "sat");
+    }
+}